requirements.txt
autocorrect==2.6.1
Flask==2.2.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
itsdangerous==2.1.2
keras==2.11.0
nltk==3.8.1
numpy==1.24.2
Pillow==9.5.0
WTForms==3.0.1
pymysql==1.0.3
tensorflow==2.11.0
email_validator==1.3.1
cryptography==40.0.1
mysql-connector-python==8.0.32